refactored threading-related classes, code broken
