+++ .gitignore
@@ -8,3 +8,12 @@ main/credentials.h
 !.vscode/tasks.json
 !.vscode/extensions.json
 *.code-workspace
+
+.DS_Store
+Thumbs.db
+.autotools
+.built
+.context
+.cproject
+.idea/*
+.vagrant
