+++ src/concurrency/Thread.cpp
@@ -1,0 +1,16 @@
+#include "Thread.h"
+
+namespace concurrency {
+
+void Thread::start(const char *name, size_t stackSize, uint32_t priority)
+{
+    auto r = xTaskCreate(callRun, name, stackSize, this, priority, &taskHandle);
+    assert(r == pdPASS);
+}
+
+void Thread::callRun(void *_this)
+{
+    ((Thread *)_this)->doRun();
+}
+
+} // namespace concurrency
